extract use-vol-status to custom hook
