+++ src/pages/controller.tsx
@@ -1,8 +1,6 @@
-import { useAtomDevtools } from 'jotai-devtools'
-import { useImmerAtom } from 'jotai-immer'
 import { Volume, VolumeOff } from 'lucide-react'
-import { useCallback, useEffect } from 'react'
-import type { GetWsMessage, VolStatus } from '../api/types'
+import { useCallback } from 'react'
+import { useVolStatus } from '../api/use-vol-status'
 import { useWebSocketApi } from '../api/use-web-socket-api'
 import { Layout } from '../components/layout'
 import { Slider } from '../components/slider'
@@ -10,59 +8,10 @@ import { Toggle } from '../components/toggle'
 import { Small } from '../components/typography'
 import { dict, MAX_VOLUME, MIN_VOLUME, testid } from '../constant'
 import { cn } from '../utils/cn'
-import { useDebounce } from '../utils/use-debounce'
-import { atom } from 'jotai'
-
-export const volStatusAtom = atom<VolStatus>()
-if (process.env.NODE_ENV !== 'production') {
-  volStatusAtom.debugLabel = 'statusAtom'
-}
-
-const useFirstLoadUpdate = (sendMessage: (message: string) => void) => {
-  const handleRefresh = useCallback(() => {
-    sendMessage(JSON.stringify({ action: 'BroadcastStatus' }))
-  }, [sendMessage])
-
-  useEffect(() => {
-    handleRefresh()
-  }, [handleRefresh])
-}
 
 export const ControllerOutput: React.FC = () => {
-  const [volStatus, updateVolStatus] = useImmerAtom(volStatusAtom)
-  useAtomDevtools(volStatusAtom)
-  const { lastMessage, sendMessage } = useWebSocketApi()
-  const debouncedVolStatus = useDebounce(volStatus, 100)
-
-  useFirstLoadUpdate(sendMessage)
-
-  // It updates status when backend server broadcasts new state
-  useEffect(() => {
-    if (lastMessage && typeof lastMessage.data === 'string') {
-      const incomeMessage = JSON.parse(lastMessage.data) as GetWsMessage
-      updateVolStatus(draft => {
-        if (incomeMessage.action === 'BroadcastStatus' && !!incomeMessage.payload) {
-          return incomeMessage.payload
-        }
-        return draft
-      })
-    }
-  }, [lastMessage, updateVolStatus])
-
-  // Send message to websocket, when debouncedVolStatus changes
-  useEffect(() => {
-    if (!debouncedVolStatus) {
-      return
-    }
-    for (const output of debouncedVolStatus.outputs) {
-      sendMessage(
-        JSON.stringify({
-          action: 'SetSink',
-          payload: { name: output.name, volume: output.volume, muted: output.muted },
-        }),
-      )
-    }
-  }, [debouncedVolStatus, sendMessage])
+  const { volStatus, updateVolStatus } = useVolStatus()
+  const { sendMessage } = useWebSocketApi()
 
   const handleVolumeChange = useCallback(
     (name: string) => (newValue: number[]) => {
